+++ compute_residual.h
@@ -1,4 +1,4 @@
 #ifndef COMPUTE_RESIDUAL_H
 #define COMPUTE_RESIDUAL_H
-int compute_residual(const int n, const double * const v1, const double * const v2, double * const residual);
+int compute_residual(const int n, const float * const v1, const float * const v2, float * const residual);
 #endif
