+++ waxpby.h
@@ -1,4 +1,4 @@
 #ifndef WAXBY_H
 #define WAXBY_H
-int waxpby (const int n, const double * const x, const double beta, const double * const y, double * const w);
+int waxpby (const int n, const float * const x, const float beta, const float * const y, float * const w);
 #endif
